Set contract country from tenant billing profile
