--- packages/server/events-processing-platform/domain/contract/command/rollout_renewal_opportunity_on_expiration_command.go
@@ -1,17 +1,0 @@
-package command
-
-import (
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/eventstore"
-)
-
-type RolloutRenewalOpportunityOnExpirationCommand struct {
-	eventstore.BaseCommand
-	AppSource string
-}
-
-func NewRolloutRenewalOpportunityOnExpirationCommand(contractId, tenant, loggedInUserId, appSource string) *RolloutRenewalOpportunityOnExpirationCommand {
-	return &RolloutRenewalOpportunityOnExpirationCommand{
-		BaseCommand: eventstore.NewBaseCommand(contractId, tenant, loggedInUserId),
-		AppSource:   appSource,
-	}
-}
