--- packages/server/events-processing-platform/domain/contract/command_handler/rollout_renewal_opportunity_on_expiration_handler.go
@@ -1,84 +1,0 @@
-package command_handler
-
-import (
-	"context"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/utils"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/config"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/domain/contract/aggregate"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/domain/contract/command"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/eventstore"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/logger"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/tracing"
-	"github.com/opentracing/opentracing-go"
-	"github.com/opentracing/opentracing-go/log"
-	"time"
-)
-
-type RolloutRenewalOpportunityOnExpirationCommandHandler interface {
-	Handle(ctx context.Context, cmd *command.RolloutRenewalOpportunityOnExpirationCommand) error
-}
-
-type rolloutRenewalOpportunityOnExpirationCommandHandler struct {
-	log logger.Logger
-	es  eventstore.AggregateStore
-	cfg config.Utils
-}
-
-func NewRolloutRenewalOpportunityOnExpirationCommandHandler(log logger.Logger, es eventstore.AggregateStore, cfg config.Utils) RolloutRenewalOpportunityOnExpirationCommandHandler {
-	return &rolloutRenewalOpportunityOnExpirationCommandHandler{
-		log: log,
-		es:  es,
-		cfg: cfg,
-	}
-}
-
-func (h *rolloutRenewalOpportunityOnExpirationCommandHandler) Handle(ctx context.Context, cmd *command.RolloutRenewalOpportunityOnExpirationCommand) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "RolloutRenewalOpportunityOnExpirationCommandHandler.Handle")
-	defer span.Finish()
-	tracing.SetCommandHandlerSpanTags(ctx, span, cmd.Tenant, cmd.LoggedInUserId)
-	span.LogFields(log.Object("command", cmd))
-
-	for attempt := 0; attempt == 0 || attempt < h.cfg.RetriesOnOptimisticLockException; attempt++ {
-		// Load or initialize the contract aggregate
-		contractAggregate, err := aggregate.LoadContractAggregate(ctx, h.es, cmd.Tenant, cmd.GetObjectID(), eventstore.LoadAggregateOptions{})
-		if err != nil {
-			tracing.TraceErr(span, err)
-			return err
-		}
-
-		if eventstore.IsAggregateNotFound(contractAggregate) {
-			tracing.TraceErr(span, eventstore.ErrAggregateNotFound)
-			return eventstore.ErrAggregateNotFound
-		}
-
-		// Apply the command to the aggregate
-		if err = contractAggregate.HandleCommand(ctx, cmd); err != nil {
-			tracing.TraceErr(span, err)
-			return err
-		}
-
-		// Persist the changes to the event store
-		err = h.es.Save(ctx, contractAggregate)
-		if err == nil {
-			return nil // Save successful
-		}
-
-		if eventstore.IsEventStoreErrorCodeWrongExpectedVersion(err) {
-			if attempt == h.cfg.RetriesOnOptimisticLockException-1 {
-				// If we have reached the maximum number of retries, return an error
-				tracing.TraceErr(span, err)
-				return err
-			}
-			// Handle concurrency error
-			span.LogFields(log.Int("retryAttempt", attempt+1))
-			time.Sleep(utils.BackOffExponentialDelay(attempt)) // backoffDelay is a function that increases the delay with each attempt
-			continue                                           // Retry
-		} else {
-			// Some other error occurred
-			tracing.TraceErr(span, err)
-			return err
-		}
-	}
-
-	return nil
-}
