+++ packages/server/events-processing-platform/service/services.go
@@ -64,12 +64,12 @@ func InitServices(cfg *config.Config, repositories *repository.Repositories, agg
 	services.IssueService = NewIssueService(log, commandHandlers.Issue)
 	services.CommentService = NewCommentService(&services, log, aggregateStore, cfg)
 	services.OpportunityService = NewOpportunityService(log, commandHandlers.Opportunity, aggregateStore)
-	services.ContractService = NewContractService(log, commandHandlers.Contract, aggregateStore, cfg)
+	services.ContractService = NewContractService(log, aggregateStore, &services)
 	services.ServiceLineItemService = NewServiceLineItemService(log, commandHandlers.ServiceLineItem, aggregateStore)
 	services.MasterPlanService = NewMasterPlanService(log, commandHandlers.MasterPlan, aggregateStore)
 	services.OrganizationPlanService = NewOrganizationPlanService(log, commandHandlers.OrganizationPlan, aggregateStore)
 	services.InvoicingCycleService = NewInvoicingCycleService(log, commandHandlers.InvoicingCycle, aggregateStore)
-	services.InvoiceService = NewInvoiceService(&services, log, aggregateStore, cfg, repositories.InvoiceRepository)
+	services.InvoiceService = NewInvoiceService(&services, log, aggregateStore, repositories.InvoiceRepository)
 	services.TenantService = NewTenantService(&services, log, aggregateStore, cfg)
 	services.CountryService = NewCountryService(&services, log, aggregateStore, cfg)
 	services.ReminderService = NewReminderService(log, aggregateStore, cfg, ebs)
